pyproject.toml
[tool.poetry]
name = "Project DRF"
version = "0.1.0"
description = ""
authors = ["Sem <semseveryanin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
ipython = "^8.25.0"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
